requirements.txt
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
attrs==25.3.0
cattrs==25.1.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-anthropic==0.3.15
langchain-core==0.3.66
langchain-mcp-adapters==0.1.7
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.4.1
lsprotocol==2025.0.0
mcp==1.9.4
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pydantic==2.11.7
pydantic-settings==2.10.0
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.3.6
starlette==0.47.1
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
xxhash==3.5.0
zstandard==0.23.0
